rx_bloc_list - Remove cupertino dependency from models, update example app
