--- packages/rx_bloc_list/example/android/app/src/main/java/com/primeholding/example/MainActivity.java
@@ -1,6 +1,0 @@
-package com.primeholding.example;
-
-import io.flutter.embedding.android.FlutterActivity;
-
-public class MainActivity extends FlutterActivity {
-}
