+++ packages/rx_bloc_list/example/android/app/src/main/kotlin/com/primeholding/example/MainActivity.kt
@@ -1,0 +1,5 @@
+package com.primeholding.example
+
+import io.flutter.embedding.android.FlutterActivity
+
+class MainActivity: FlutterActivity()
