+++ packages/rx_bloc_list/example/pubspec.yaml
@@ -14,7 +14,6 @@ dependencies:
   flutter_rx_bloc:
   rx_bloc:
   rx_bloc_list:
-    path: ..
   rxdart: ^0.28.0
 
 dev_dependencies:
@@ -23,14 +22,6 @@ dev_dependencies:
   flutter_test:
     sdk: flutter
   rx_bloc_generator:
-
-
-dependency_overrides:
-  flutter_rx_bloc:
-    path: ../../flutter_rx_bloc
-  rx_bloc:
-    path: ../../rx_bloc
-
 
 flutter:
   uses-material-design: true
