+++ packages/rx_bloc_list/example/pubspec_overrides.yaml
@@ -1,0 +1,7 @@
+dependency_overrides:
+  flutter_rx_bloc:
+    path: ../../flutter_rx_bloc
+  rx_bloc:
+    path: ../../rx_bloc
+  rx_bloc_list:
+    path: ..
