+++ packages/rx_bloc_list/lib/extensions.dart
@@ -1,0 +1,4 @@
+library extensions;
+
+export 'src/extensions/identifiable_extensions.dart';
+export 'src/extensions/paginated_list_extensions.dart';
