+++ packages/rx_bloc_list/lib/rx_bloc_list.dart
@@ -1,4 +1,5 @@
 library rx_bloc_list;
 
+export 'extensions.dart';
 export 'models.dart';
 export 'widgets.dart';
