+++ packages/rx_bloc_list/pubspec.yaml
@@ -1,6 +1,6 @@
 name: rx_bloc_list
 description: The rx_bloc_list package facilitates implementing infinity scroll and pull-to-refresh features with minimal setup. This package is meant to be used along with RxBloc ecosystem.
-version: 5.0.0
+version: 5.0.1
 homepage: https://www.primeholding.com
 repository: https://github.com/Prime-Holding/rx_bloc/tree/master/packages/rx_bloc_list
 issue_tracker: https://github.com/Prime-Holding/rx_bloc/issues
