+++ packages/rx_bloc_list/test/paginated_list_extensions_test.dart
@@ -1,6 +1,7 @@
 import 'package:flutter/widgets.dart';
 import 'package:flutter_test/flutter_test.dart';
 import 'package:rx_bloc/rx_bloc.dart';
+import 'package:rx_bloc_list/extensions.dart';
 import 'package:rx_bloc_list/models.dart';
 import 'package:rxdart/rxdart.dart';
